--- src/main/java/com/educator/core/answer_session/enums/SubjectColor.java
@@ -1,7 +1,0 @@
-package com.educator.core.answer_session.enums;
-
-public enum SubjectColor {
-
-    GREEN,
-    YELLOW
-}
