+++ src/main/java/com/educator/core/subject/SubjectColor.java
@@ -1,0 +1,6 @@
+package com.educator.core.subject;
+
+public enum SubjectColor {
+    TO_SEVEN_DAYS,
+    AFTER_SEVEN_DAYS
+}
