+++ src/main/java/com/educator/core/subject/SubjectService.java
@@ -1,6 +1,15 @@
 package com.educator.core.subject;
+import com.educator.core.answer_session.AnswerSession;
+import com.educator.core.answer_session.AnswerSessionRepository;
+import com.educator.core.answer_session.enums.StatusAnswerSession;
+import com.educator.core.subject.dto.CheckCompletedSessionsDto;
+import com.educator.core.subject.dto.SubjectDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.time.LocalDate;
+import java.time.temporal.ChronoUnit;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -8,7 +17,13 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class SubjectService {
 
+    private static final double BORDER_FINISHED_SUBJECT = 0.8;
+
+    private static final int BORDER_SEVEN_DAYS = 7;
+
     private final SubjectRepository subjectRepository;
+
+    private final AnswerSessionRepository answerSessionRepository;
 
     private final SubjectMapper subjectMapper;
   
@@ -32,4 +47,37 @@ public class SubjectService {
         return subjectMapper.mapToDtoSubjectList(subjectRepository.findAll()).stream()
                 .filter(i -> i.getCourseId().equals(courseId)).collect(Collectors.toList());
     }
+
+    public List<CheckCompletedSessionsDto> getAllNumbersOfCorrectAnswersAtLeast80Percent(Long courseId) {
+        List<Long> subjectsIdFilterByCourseId = subjectRepository.findByCourseId(courseId).stream().map(Subject::getId).collect(Collectors.toList());
+        List<AnswerSession> answerSessionsCompletedList = answerSessionRepository.findByStatusAnswerSession(StatusAnswerSession.COMPLETED);
+        return getCheckCompletedSessionsDtos(subjectsIdFilterByCourseId, answerSessionsCompletedList);
+
+    }
+
+    private static List<CheckCompletedSessionsDto> getCheckCompletedSessionsDtos(List<Long> subjectsIdFilterByCourseId, List<AnswerSession> answerSessionsCompletedList) {
+        List<CheckCompletedSessionsDto> checkCompletedSessionsDtoList = new ArrayList<>();
+        for (AnswerSession answerSessionCompleted : answerSessionsCompletedList) {
+            double checkAllAndCorrectAnswers = getCheckAllAndCorrectAnswers(answerSessionCompleted);
+            if (checkAllAndCorrectAnswers >= BORDER_FINISHED_SUBJECT && subjectsIdFilterByCourseId.contains(answerSessionCompleted.getSubject().getId())) {
+                if (getDifferenceInDaysBetweenTheStartedSessionAndToday(answerSessionCompleted) <= BORDER_SEVEN_DAYS) {
+                    checkCompletedSessionsDtoList.add(new CheckCompletedSessionsDto(answerSessionCompleted.getId(), answerSessionCompleted.getSubject().getId(), SubjectColor.TO_SEVEN_DAYS));
+                } else {
+                    checkCompletedSessionsDtoList.add(new CheckCompletedSessionsDto(answerSessionCompleted.getId(), answerSessionCompleted.getSubject().getId(), SubjectColor.AFTER_SEVEN_DAYS));
+                }
+            }
+        }
+        return checkCompletedSessionsDtoList;
+    }
+
+    private static double getCheckAllAndCorrectAnswers(AnswerSession answerSessionCompleted) {
+        return answerSessionCompleted.getAllAnswers() == 0 ? 0: (double) answerSessionCompleted.getCorrectAnswers() / answerSessionCompleted.getAllAnswers();
+    }
+
+    private static long getDifferenceInDaysBetweenTheStartedSessionAndToday(AnswerSession answerSessionCompleted) {
+        long differenceInDaysBetweenTheStartedSessionAndToday;
+        differenceInDaysBetweenTheStartedSessionAndToday = ChronoUnit.DAYS.between(answerSessionCompleted.getSessionStartDate(), LocalDate.now());
+        return differenceInDaysBetweenTheStartedSessionAndToday;
+    }
 }
+
